refactor(api): move reset password service
