+++ api/tests/identity-access-management/integration/domain/services/reset-password.service.test.js
@@ -1,10 +1,10 @@
-import * as resetPasswordService from '../../../../lib/domain/services/reset-password-service.js';
-import { expect } from '../../../test-helper.js';
+import * as resetPasswordService from '../../../../../src/identity-access-management/domain/services/reset-password.service.js';
+import { expect } from '../../../../test-helper.js';
 
-describe('Integration | Service | Password Service', function () {
+describe('Integration | Identity Access Management | Domain | Service | reset-password', function () {
   describe('#generateTemporaryKey', function () {
     context('when two users send a request at the same second', function () {
-      it('should generate different temporaryKeys', function () {
+      it('generates different temporaryKeys', function () {
         // when
         const temporaryKeyUser1 = resetPasswordService.generateTemporaryKey();
         const temporaryKeyUser2 = resetPasswordService.generateTemporaryKey();
