+++ api/tests/integration/domain/usecases/create-password-reset-demand_test.js
@@ -1,7 +1,7 @@
 import { UserNotFoundError } from '../../../../lib/domain/errors.js';
 import * as mailService from '../../../../lib/domain/services/mail-service.js';
-import * as resetPasswordService from '../../../../lib/domain/services/reset-password-service.js';
 import { createPasswordResetDemand } from '../../../../lib/domain/usecases/create-password-reset-demand.js';
+import * as resetPasswordService from '../../../../src/identity-access-management/domain/services/reset-password.service.js';
 import * as resetPasswordDemandRepository from '../../../../src/identity-access-management/infrastructure/repositories/reset-password-demand.repository.js';
 import * as userRepository from '../../../../src/shared/infrastructure/repositories/user-repository.js';
 import { catchErr, databaseBuilder, expect } from '../../../test-helper.js';
