+++ api/tests/integration/domain/usecases/get-user-by-reset-password-demand_test.js
@@ -1,6 +1,6 @@
 import { PasswordResetDemandNotFoundError, UserNotFoundError } from '../../../../lib/domain/errors.js';
-import * as resetPasswordService from '../../../../lib/domain/services/reset-password-service.js';
 import { getUserByResetPasswordDemand } from '../../../../lib/domain/usecases/get-user-by-reset-password-demand.js';
+import * as resetPasswordService from '../../../../src/identity-access-management/domain/services/reset-password.service.js';
 import { InvalidTemporaryKeyError } from '../../../../src/shared/domain/errors.js';
 import { User } from '../../../../src/shared/domain/models/User.js';
 import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
